+++ docs/changelog.rst
@@ -29,6 +29,9 @@ v0.1
   `#4 <https://github.com/msoeken/kitty/pull/4>`_
   `#8 <https://github.com/msoeken/kitty/pull/8>`_
 
+* Predicates: ``has_var``
+  `#11 <https://github.com/msoeken/kitty/pull/11>`_
+
 * Operators: ``~``, ``==``, ``<``
   `#8 <https://github.com/msoeken/kitty/pull/8>`_